fix(ui): use UINT16 width/advance for 256px art glyph table
SFB_FONT_GLYPH stores width/advance as UINT8, which cannot hold the
256px art cell size and broke the build (-Wconstant-conversion). Add a
dedicated SFB_ART_GLYPH struct with UINT16 metrics for the art font.

+++ submodules/uefi/edk2/QcomModulePkg/Application/LinuxLoader/SuperFbFont.c
@@ -36,8 +36,8 @@ SfbFontGetGlyph (IN CHAR16 Ch,
 BOOLEAN
 SfbArtGetGlyph (IN CHAR16 Ch,
                 OUT UINT32 *Offset,
-                OUT UINT8  *Width,
-                OUT UINT8  *Advance)
+                OUT UINT16 *Width,
+                OUT UINT16 *Advance)
 {
   UINTN  Index;
